[#858] Move Test_Resource_Replication_With_Retry to timing_test suite
